use `f64` for price (#159)

+++ bitskins/src/http.rs
@@ -235,7 +235,7 @@ impl HttpClient {
             .balance as f64)
     }
 
-    pub async fn buy_item(&self, item_id: &str, price: i32) -> Result<()> {
+    pub async fn buy_item(&self, item_id: &str, price: f64) -> Result<()> {
         self.post(
             Endpoint::BuySingle,
             json!({
